Reviewed: minimal, correct fix — cast to int at all three broken call sites, consistent approach across all of them. Independently swept the whole Services/ tree myself and confirmed no other .Filter call passes a long (season/week are int everywhere, the List-based position_id filter in PlayerRepository was never affected). Ran dotnet test myself: 84/84 pass. CI green.

+++ Services/Repositories/RunRepository.cs
@@ -27,7 +27,9 @@ public class RunRepository : IRunRepository
             .Select("payload_hash")
             .Filter("season", Constants.Operator.Equals, season)
             .Filter("week", Constants.Operator.Equals, week)
-            .Filter("position_id", Constants.Operator.Equals, positionId)
+            // Postgrest's Filter(...) only accepts string/int/float/List/Dictionary/FullTextSearchConfig/Range
+            // criterion types (not long) - cast down; position ids are always small (1-5).
+            .Filter("position_id", Constants.Operator.Equals, (int)positionId)
             .Filter("scoring", Constants.Operator.Equals, scoring)
             .Order("fetched_at", Constants.Ordering.Descending)
             .Limit(1)
